requirements.txt
Django==4.2
semgrep==1.104.0
djangorestframework==3.12.4
drf-yasg==1.20.0 